requirements.txt
numpy>=1.26,<2
Pillow>=10.0
protobuf>=4.25,<6.0
torch>=2.2
pytest>=8.0
